SNOW-3396010: Split platform_detection into smaller files
