+++ sf_core/src/telemetry/mod.rs
@@ -7,7 +7,6 @@ pub mod serialization;
 #[doc(hidden)]
 pub mod snowflake_exporter;
 
-pub(crate) mod aws_identity;
 pub mod os_details;
 pub mod platform_detection;
